Lab02/requirements.txt
pyTelegramBotAPI==4.32.0
scikit-learn==1.8.0
matplotlib==3.10.8
wordcloud==1.9.6
python-dotenv==1.2.2
numpy==1.26.4
scipy==1.17.1
gensim==4.3.3
datasets==3.6.0
